move studies chart into it's own file
